Fixes issue #493 extract ascii art logic (#561)
* added asciiartmanager class and asciiartmanagertest. changed buildservice.java and commandlinerunner.java to utilize the new class

* Extracted AsciiArtManager logic into separate class. Extracted AsciiArt testing logic into separate class. Fixes #493.

* Fix issue in CommandLineRunner.java
